+++ app/lib/frontend/templates/views/pkg/tags.dart
@@ -1,0 +1,143 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import '../../../dom/dom.dart' as d;
+
+class SimpleTag {
+  final String status;
+  final String text;
+  final String title;
+  final String? href;
+
+  bool get hasHref => href != null;
+
+  SimpleTag({
+    required this.status,
+    required this.text,
+    required this.title,
+    this.href,
+  });
+
+  factory SimpleTag.discontinued() {
+    return SimpleTag(
+      status: 'discontinued',
+      text: 'discontinued',
+      title: 'Package was discontinued.',
+    );
+  }
+
+  factory SimpleTag.unlisted() {
+    return SimpleTag(
+      status: 'unlisted',
+      text: 'unlisted',
+      title: 'Package is unlisted, this means that while the package is still '
+          'publicly available the author has decided that it should not appear '
+          'in search results with default search filters. This is typically '
+          'done because this package is meant to support another package, '
+          'rather than being consumed directly.',
+    );
+  }
+
+  factory SimpleTag.obsolete() {
+    return SimpleTag(
+      status: 'missing',
+      text: 'outdated',
+      title: 'Package version too old, check latest stable.',
+    );
+  }
+
+  factory SimpleTag.legacy() {
+    return SimpleTag(
+      status: 'legacy',
+      text: 'Dart 2 incompatible',
+      title: 'Package does not support Dart 2.',
+    );
+  }
+
+  factory SimpleTag.awaiting() {
+    return SimpleTag(
+      status: 'missing',
+      text: '[awaiting]',
+      title: 'Analysis should be ready soon.',
+    );
+  }
+
+  factory SimpleTag.unidentified({String? href}) {
+    return SimpleTag(
+      status: 'unidentified',
+      text: '[unidentified]',
+      title: 'Check the scores tab for further details.',
+      href: href,
+    );
+  }
+}
+
+class BadgeTag {
+  final String sdk;
+  final String title;
+  final String href;
+  final List<BadgeSubTag> subTags;
+
+  BadgeTag({
+    required this.sdk,
+    required this.title,
+    required this.href,
+    required this.subTags,
+  });
+}
+
+class BadgeSubTag {
+  final String text;
+  final String title;
+  final String href;
+
+  BadgeSubTag({
+    required this.text,
+    required this.title,
+    required this.href,
+  });
+}
+
+d.Node tagsNode({
+  required List<SimpleTag> simpleTags,
+  required List<BadgeTag> badgeTags,
+}) {
+  return d.fragment([
+    ...badgeTags.map(
+      (t) => d.div(
+        classes: ['-pub-tag-badge'],
+        children: [
+          d.a(
+            classes: ['tag-badge-main'],
+            title: t.title,
+            href: t.href,
+            text: t.sdk,
+          ),
+          ...t.subTags.map(
+            (s) => d.a(
+              classes: ['tag-badge-sub'],
+              title: s.title,
+              href: s.href,
+              text: s.text,
+            ),
+          ),
+        ],
+      ),
+    ),
+    ...simpleTags.map(
+      (t) => t.hasHref
+          ? d.a(
+              classes: ['package-tag', t.status],
+              href: t.href!,
+              title: t.title,
+              text: t.text,
+            )
+          : d.span(
+              classes: ['package-tag', t.status],
+              attributes: {'title': t.title},
+              text: t.text,
+            ),
+    ),
+  ]);
+}